% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.s3_operations.R
\name{delete_bucket}
\alias{delete_bucket}
\title{Deletes the bucket}
\usage{
delete_bucket(
  Bucket = "string"
)
}
\arguments{
\item{Bucket}{[required]}
}
\description{
Deletes the bucket. All objects (including all object versions and Delete Markers) in the bucket must be deleted before the bucket itself can be deleted.
}
\examples{
# The following example deletes the specified bucket.
delete_bucket(
  Bucket = "forrandall2"
)

}
