% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.s3_operations.R
\name{delete_bucket_cors}
\alias{delete_bucket_cors}
\title{Deletes the CORS configuration information set for the bucket}
\usage{
delete_bucket_cors(
  Bucket = "string"
)
}
\arguments{
\item{Bucket}{[required]}
}
\description{
Deletes the CORS configuration information set for the bucket.
}
\examples{
# The following example deletes CORS configuration on a bucket.
delete_bucket_cors(
  Bucket = "examplebucket"
)

}
