% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.s3_operations.R
\name{delete_bucket_inventory_configuration}
\alias{delete_bucket_inventory_configuration}
\title{Deletes an inventory configuration (identified by the inventory ID) from the bucket}
\usage{
delete_bucket_inventory_configuration(
  Bucket = "string",
  Id = "string"
)
}
\arguments{
\item{Bucket}{[required] The name of the bucket containing the inventory configuration to delete.}

\item{Id}{[required] The ID used to identify the inventory configuration.}
}
\description{
Deletes an inventory configuration (identified by the inventory ID) from the bucket.
}
\examples{

}
