% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.s3_operations.R
\name{delete_bucket_policy}
\alias{delete_bucket_policy}
\title{Deletes the policy from the bucket}
\usage{
delete_bucket_policy(Bucket)
}
\arguments{
\item{Bucket}{[required]}
}
\description{
Deletes the policy from the bucket.
}
\section{Accepted Parameters}{
\preformatted{delete_bucket_policy(
  Bucket = "string"
)
}
}

\examples{
# The following example deletes bucket policy on the specified bucket.
\donttest{delete_bucket_policy(
  Bucket = "examplebucket"
)}

}
