% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.s3_operations.R
\name{delete_bucket_replication}
\alias{delete_bucket_replication}
\title{Deletes the replication configuration from the bucket}
\usage{
delete_bucket_replication(Bucket)
}
\arguments{
\item{Bucket}{[required] Deletes the replication subresource associated with the specified bucket.

There is usually some time lag before replication configuration deletion is fully propagated to all the Amazon S3 systems.

For more information, see \href{%20https://docs.aws.amazon.com/AmazonS3/latest/dev/crr.html}{Cross-Region Replication (CRR)} in the Amazon S3 Developer Guide.}
}
\description{
Deletes the replication configuration from the bucket.
}
\section{Accepted Parameters}{
\preformatted{delete_bucket_replication(
  Bucket = "string"
)
}
}

\examples{
# The following example deletes replication configuration set on bucket.
\donttest{delete_bucket_replication(
  Bucket = "example"
)}

}
