% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.s3_operations.R
\name{delete_object_tagging}
\alias{delete_object_tagging}
\title{Removes the tag-set from an existing object}
\usage{
delete_object_tagging(Bucket, Key, VersionId = NULL)
}
\arguments{
\item{Bucket}{[required]}

\item{Key}{[required]}

\item{VersionId}{The versionId of the object that the tag-set will be removed from.}
}
\description{
Removes the tag-set from an existing object.
}
\section{Accepted Parameters}{
\preformatted{delete_object_tagging(
  Bucket = "string",
  Key = "string",
  VersionId = "string"
)
}
}

\examples{
# The following example removes tag set associated with the specified
# object version. The request specifies both the object key and object
# version.
\donttest{delete_object_tagging(
  Bucket = "examplebucket",
  Key = "HappyFace.jpg",
  VersionId = "ydlaNkwWm0SfKJR.T1b1fIdPRbldTYRI"
)}

# The following example removes tag set associated with the specified
# object. If the bucket is versioning enabled, the operation removes tag
# set from the latest object version.
\donttest{delete_object_tagging(
  Bucket = "examplebucket",
  Key = "HappyFace.jpg"
)}

}
