% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.s3_operations.R
\name{get_bucket_accelerate_configuration}
\alias{get_bucket_accelerate_configuration}
\title{Returns the accelerate configuration of a bucket}
\usage{
get_bucket_accelerate_configuration(
  Bucket = "string"
)
}
\arguments{
\item{Bucket}{[required] Name of the bucket for which the accelerate configuration is retrieved.}
}
\description{
Returns the accelerate configuration of a bucket.
}
\examples{

}
