% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.s3_operations.R
\name{get_bucket_acl}
\alias{get_bucket_acl}
\title{Gets the access control policy for the bucket}
\usage{
get_bucket_acl(
  Bucket = "string"
)
}
\arguments{
\item{Bucket}{[required]}
}
\description{
Gets the access control policy for the bucket.
}
\examples{

}
