% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.s3_operations.R
\name{get_bucket_analytics_configuration}
\alias{get_bucket_analytics_configuration}
\title{Gets an analytics configuration for the bucket (specified by the analytics configuration ID)}
\usage{
get_bucket_analytics_configuration(
  Bucket = "string",
  Id = "string"
)
}
\arguments{
\item{Bucket}{[required] The name of the bucket from which an analytics configuration is retrieved.}

\item{Id}{[required] The identifier used to represent an analytics configuration.}
}
\description{
Gets an analytics configuration for the bucket (specified by the analytics configuration ID).
}
\examples{

}
