% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.s3_operations.R
\name{get_bucket_cors}
\alias{get_bucket_cors}
\title{Returns the CORS configuration for the bucket}
\usage{
get_bucket_cors(
  Bucket = "string"
)
}
\arguments{
\item{Bucket}{[required]}
}
\description{
Returns the CORS configuration for the bucket.
}
\examples{
# The following example returns cross-origin resource sharing (CORS)
# configuration set on a bucket.
get_bucket_cors(
  Bucket = "examplebucket"
)

}
