% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.s3_operations.R
\name{get_bucket_encryption}
\alias{get_bucket_encryption}
\title{Returns the server-side encryption configuration of a bucket}
\usage{
get_bucket_encryption(
  Bucket = "string"
)
}
\arguments{
\item{Bucket}{[required] The name of the bucket from which the server-side encryption configuration is retrieved.}
}
\description{
Returns the server-side encryption configuration of a bucket.
}
\examples{

}
