% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.s3_operations.R
\name{get_bucket_lifecycle_configuration}
\alias{get_bucket_lifecycle_configuration}
\title{Returns the lifecycle configuration information set on the bucket}
\usage{
get_bucket_lifecycle_configuration(
  Bucket = "string"
)
}
\arguments{
\item{Bucket}{[required]}
}
\description{
Returns the lifecycle configuration information set on the bucket.
}
\examples{
# The following example retrieves lifecycle configuration on set on a
# bucket. 
get_bucket_lifecycle_configuration(
  Bucket = "examplebucket"
)

}
