% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.s3_operations.R
\name{get_bucket_location}
\alias{get_bucket_location}
\title{Returns the region the bucket resides in}
\usage{
get_bucket_location(Bucket)
}
\arguments{
\item{Bucket}{[required]}
}
\description{
Returns the region the bucket resides in.
}
\section{Accepted Parameters}{
\preformatted{get_bucket_location(
  Bucket = "string"
)
}
}

\examples{
# The following example returns bucket location.
\donttest{get_bucket_location(
  Bucket = "examplebucket"
)}

}
