% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.s3_operations.R
\name{get_bucket_logging}
\alias{get_bucket_logging}
\title{Returns the logging status of a bucket and the permissions users have to view and modify that status}
\usage{
get_bucket_logging(
  Bucket = "string"
)
}
\arguments{
\item{Bucket}{[required]}
}
\description{
Returns the logging status of a bucket and the permissions users have to view and modify that status. To use GET, you must be the bucket owner.
}
\examples{

}
