% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.s3_operations.R
\name{get_bucket_notification}
\alias{get_bucket_notification}
\title{Deprecated, see the GetBucketNotificationConfiguration operation}
\usage{
get_bucket_notification(
  Bucket = "string"
)
}
\arguments{
\item{Bucket}{[required] Name of the bucket to get the notification configuration for.}
}
\description{
Deprecated, see the GetBucketNotificationConfiguration operation.
}
\examples{
# The following example returns notification configuration set on a
# bucket.
get_bucket_notification(
  Bucket = "examplebucket"
)

# The following example returns notification configuration set on a
# bucket.
get_bucket_notification(
  Bucket = "examplebucket"
)

}
