% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.s3_operations.R
\name{get_bucket_notification_configuration}
\alias{get_bucket_notification_configuration}
\title{Returns the notification configuration of a bucket}
\usage{
get_bucket_notification_configuration(
  Bucket = "string"
)
}
\arguments{
\item{Bucket}{[required] Name of the bucket to get the notification configuration for.}
}
\description{
Returns the notification configuration of a bucket.
}
\examples{

}
