% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.s3_operations.R
\name{get_bucket_policy}
\alias{get_bucket_policy}
\title{Returns the policy of a specified bucket}
\usage{
get_bucket_policy(Bucket)
}
\arguments{
\item{Bucket}{[required]}
}
\description{
Returns the policy of a specified bucket.
}
\section{Accepted Parameters}{
\preformatted{get_bucket_policy(
  Bucket = "string"
)
}
}

\examples{
# The following example returns bucket policy associated with a bucket.
\donttest{get_bucket_policy(
  Bucket = "examplebucket"
)}

}
