% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.s3_operations.R
\name{get_bucket_policy_status}
\alias{get_bucket_policy_status}
\title{Retrieves the policy status for an Amazon S3 bucket, indicating whether the bucket is public}
\usage{
get_bucket_policy_status(Bucket)
}
\arguments{
\item{Bucket}{[required] The name of the Amazon S3 bucket whose policy status you want to retrieve.}
}
\description{
Retrieves the policy status for an Amazon S3 bucket, indicating whether the bucket is public.
}
\section{Accepted Parameters}{
\preformatted{get_bucket_policy_status(
  Bucket = "string"
)
}
}

