% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.s3_operations.R
\name{get_bucket_request_payment}
\alias{get_bucket_request_payment}
\title{Returns the request payment configuration of a bucket}
\usage{
get_bucket_request_payment(
  Bucket = "string"
)
}
\arguments{
\item{Bucket}{[required]}
}
\description{
Returns the request payment configuration of a bucket.
}
\examples{
# The following example retrieves bucket versioning configuration.
get_bucket_request_payment(
  Bucket = "examplebucket"
)

}
