% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.s3_operations.R
\name{get_bucket_versioning}
\alias{get_bucket_versioning}
\title{Returns the versioning state of a bucket}
\usage{
get_bucket_versioning(Bucket)
}
\arguments{
\item{Bucket}{[required]}
}
\description{
Returns the versioning state of a bucket.
}
\section{Accepted Parameters}{
\preformatted{get_bucket_versioning(
  Bucket = "string"
)
}
}

\examples{
# The following example retrieves bucket versioning configuration.
\donttest{get_bucket_versioning(
  Bucket = "examplebucket"
)}

}
