% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.s3_operations.R
\name{get_object_acl}
\alias{get_object_acl}
\title{Returns the access control list (ACL) of an object}
\usage{
get_object_acl(Bucket, Key, VersionId = NULL, RequestPayer = NULL)
}
\arguments{
\item{Bucket}{[required]}

\item{Key}{[required]}

\item{VersionId}{VersionId used to reference a specific version of the object.}

\item{RequestPayer}{}
}
\description{
Returns the access control list (ACL) of an object.
}
\section{Accepted Parameters}{
\preformatted{get_object_acl(
  Bucket = "string",
  Key = "string",
  VersionId = "string",
  RequestPayer = "requester"
)
}
}

\examples{
# The following example retrieves access control list (ACL) of an object.
\donttest{get_object_acl(
  Bucket = "examplebucket",
  Key = "HappyFace.jpg"
)}

}
