% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.s3_operations.R
\name{get_object_lock_configuration}
\alias{get_object_lock_configuration}
\title{Gets the Object Lock configuration for a bucket}
\usage{
get_object_lock_configuration(
  Bucket = "string"
)
}
\arguments{
\item{Bucket}{[required] The bucket whose Object Lock configuration you want to retrieve.}
}
\description{
Gets the Object Lock configuration for a bucket. The rule specified in the Object Lock configuration will be applied by default to every new object placed in the specified bucket.
}
\examples{

}
