% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.s3_operations.R
\name{get_object_retention}
\alias{get_object_retention}
\title{Retrieves an object's retention settings}
\usage{
get_object_retention(Bucket, Key, VersionId = NULL,
  RequestPayer = NULL)
}
\arguments{
\item{Bucket}{[required] The bucket containing the object whose retention settings you want to retrieve.}

\item{Key}{[required] The key name for the object whose retention settings you want to retrieve.}

\item{VersionId}{The version ID for the object whose retention settings you want to retrieve.}

\item{RequestPayer}{}
}
\description{
Retrieves an object's retention settings.
}
\section{Accepted Parameters}{
\preformatted{get_object_retention(
  Bucket = "string",
  Key = "string",
  VersionId = "string",
  RequestPayer = "requester"
)
}
}

