% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.s3_operations.R
\name{get_object_torrent}
\alias{get_object_torrent}
\title{Return torrent files from a bucket}
\usage{
get_object_torrent(
  Bucket = "string",
  Key = "string",
  RequestPayer = "requester"
)
}
\arguments{
\item{Bucket}{[required]}

\item{Key}{[required]}

\item{RequestPayer}{}
}
\description{
Return torrent files from a bucket.
}
\examples{
# The following example retrieves torrent files of an object.
get_object_torrent(
  Bucket = "examplebucket",
  Key = "HappyFace.jpg"
)

}
