% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.s3_operations.R
\name{get_public_access_block}
\alias{get_public_access_block}
\title{Retrieves the \code{PublicAccessBlock} configuration for an Amazon S3 bucket}
\usage{
get_public_access_block(
  Bucket = "string"
)
}
\arguments{
\item{Bucket}{[required] The name of the Amazon S3 bucket whose \code{PublicAccessBlock} configuration you want to retrieve.}
}
\description{
Retrieves the \code{PublicAccessBlock} configuration for an Amazon S3 bucket.
}
\examples{

}
