% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.s3_operations.R
\name{head_bucket}
\alias{head_bucket}
\title{This operation is useful to determine if a bucket exists and you have permission to access it}
\usage{
head_bucket(Bucket)
}
\arguments{
\item{Bucket}{[required]}
}
\description{
This operation is useful to determine if a bucket exists and you have permission to access it.
}
\section{Accepted Parameters}{
\preformatted{head_bucket(
  Bucket = "string"
)
}
}

\examples{
# This operation checks to see if a bucket exists.
\donttest{head_bucket(
  Bucket = "acl1"
)}

}
