% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.s3_operations.R
\name{list_bucket_inventory_configurations}
\alias{list_bucket_inventory_configurations}
\title{Returns a list of inventory configurations for the bucket}
\usage{
list_bucket_inventory_configurations(
  Bucket = "string",
  ContinuationToken = "string"
)
}
\arguments{
\item{Bucket}{[required] The name of the bucket containing the inventory configurations to retrieve.}

\item{ContinuationToken}{The marker used to continue an inventory configuration listing that has been truncated. Use the NextContinuationToken from a previously truncated list response to continue the listing. The continuation token is an opaque value that Amazon S3 understands.}
}
\description{
Returns a list of inventory configurations for the bucket.
}
\examples{

}
