% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.s3_operations.R
\name{list_buckets}
\alias{list_buckets}
\title{Returns a list of all buckets owned by the authenticated sender of the request}
\usage{
list_buckets()
}
\description{
Returns a list of all buckets owned by the authenticated sender of the request.
}
\examples{
# The following example return versions of an object with specific key
# name prefix. The request limits the number of items returned to two. If
# there are are more than two object version, S3 returns NextToken in the
# response. You can specify this token value in your next request to fetch
# next set of object versions.
list_buckets()

}
