% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.s3_operations.R
\name{list_objects}
\alias{list_objects}
\title{Returns some or all (up to 1000) of the objects in a bucket}
\usage{
list_objects(
  Bucket = "string",
  Delimiter = "string",
  EncodingType = "url",
  Marker = "string",
  MaxKeys = 123,
  Prefix = "string",
  RequestPayer = "requester"
)
}
\arguments{
\item{Bucket}{[required]}

\item{Delimiter}{A delimiter is a character you use to group keys.}

\item{EncodingType}{}

\item{Marker}{Specifies the key to start with when listing objects in a bucket.}

\item{MaxKeys}{Sets the maximum number of keys returned in the response. The response might contain fewer keys but will never contain more.}

\item{Prefix}{Limits the response to keys that begin with the specified prefix.}

\item{RequestPayer}{Confirms that the requester knows that she or he will be charged for the list objects request. Bucket owners need not specify this parameter in their requests.}
}
\description{
Returns some or all (up to 1000) of the objects in a bucket. You can use the request parameters as selection criteria to return a subset of the objects in a bucket.
}
\examples{
# The following example list two objects in a bucket.
list_objects(
  Bucket = "examplebucket",
  MaxKeys = "2"
)

}
