% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.s3_operations.R
\name{list_parts}
\alias{list_parts}
\title{Lists the parts that have been uploaded for a specific multipart upload}
\usage{
list_parts(Bucket, Key, MaxParts = NULL, PartNumberMarker = NULL,
  UploadId, RequestPayer = NULL)
}
\arguments{
\item{Bucket}{[required]}

\item{Key}{[required]}

\item{MaxParts}{Sets the maximum number of parts to return.}

\item{PartNumberMarker}{Specifies the part after which listing should begin. Only parts with higher part numbers will be listed.}

\item{UploadId}{[required] Upload ID identifying the multipart upload whose parts are being listed.}

\item{RequestPayer}{}
}
\description{
Lists the parts that have been uploaded for a specific multipart upload.
}
\section{Accepted Parameters}{
\preformatted{list_parts(
  Bucket = "string",
  Key = "string",
  MaxParts = 123,
  PartNumberMarker = 123,
  UploadId = "string",
  RequestPayer = "requester"
)
}
}

\examples{
# The following example lists parts uploaded for a specific multipart
# upload.
\donttest{list_parts(
  Bucket = "examplebucket",
  Key = "bigobject",
  UploadId = "example7YPBOJuoFiQ9cz4P3Pe6FIZwO4f7wN93uHsNBEw97pl5eNwzExg0LAT2dUN91cOmrEQHDsP3WA60CEg--"
)}

}
