% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.s3_package.R
\docType{package}
\name{paws.s3-package}
\alias{paws.s3}
\alias{paws.s3-package}
\title{paws.s3: Amazon Simple Storage Service}
\description{

}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker \email{adam.banker39@gmail.com}
}

Other contributors:
\itemize{
  \item Amazon.com, Inc. [copyright holder]
}

}
\keyword{internal}
