% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.s3_operations.R
\name{put_bucket_accelerate_configuration}
\alias{put_bucket_accelerate_configuration}
\title{Sets the accelerate configuration of an existing bucket}
\usage{
put_bucket_accelerate_configuration(
  Bucket = "string",
  AccelerateConfiguration = list(
    Status = "Enabled"|"Suspended"
  )
)
}
\arguments{
\item{Bucket}{[required] Name of the bucket for which the accelerate configuration is set.}

\item{AccelerateConfiguration}{[required] Specifies the Accelerate Configuration you want to set for the bucket.}
}
\description{
Sets the accelerate configuration of an existing bucket.
}
\examples{

}
