% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.s3_operations.R
\name{put_bucket_encryption}
\alias{put_bucket_encryption}
\title{Creates a new server-side encryption configuration (or replaces an existing one, if present)}
\usage{
put_bucket_encryption(
  Bucket = "string",
  ContentMD5 = "string",
  ServerSideEncryptionConfiguration = list(
    Rules = list(
      list(
        ApplyServerSideEncryptionByDefault = list(
          SSEAlgorithm = "AES256"|"aws:kms",
          KMSMasterKeyID = "string"
        )
      )
    )
  )
)
}
\arguments{
\item{Bucket}{[required] The name of the bucket for which the server-side encryption configuration is set.}

\item{ContentMD5}{The base64-encoded 128-bit MD5 digest of the server-side encryption configuration.}

\item{ServerSideEncryptionConfiguration}{[required]}
}
\description{
Creates a new server-side encryption configuration (or replaces an existing one, if present).
}
\examples{

}
