% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.s3_operations.R
\name{put_bucket_lifecycle}
\alias{put_bucket_lifecycle}
\title{Deprecated, see the PutBucketLifecycleConfiguration operation}
\usage{
put_bucket_lifecycle(
  Bucket = "string",
  ContentMD5 = "string",
  LifecycleConfiguration = list(
    Rules = list(
      list(
        Expiration = list(
          Date = as.POSIXct("2015-01-01"),
          Days = 123,
          ExpiredObjectDeleteMarker = TRUE|FALSE
        ),
        ID = "string",
        Prefix = "string",
        Status = "Enabled"|"Disabled",
        Transition = list(
          Date = as.POSIXct("2015-01-01"),
          Days = 123,
          StorageClass = "GLACIER"|"STANDARD_IA"|"ONEZONE_IA"|"INTELLIGENT_TIERING"
        ),
        NoncurrentVersionTransition = list(
          NoncurrentDays = 123,
          StorageClass = "GLACIER"|"STANDARD_IA"|"ONEZONE_IA"|"INTELLIGENT_TIERING"
        ),
        NoncurrentVersionExpiration = list(
          NoncurrentDays = 123
        ),
        AbortIncompleteMultipartUpload = list(
          DaysAfterInitiation = 123
        )
      )
    )
  )
)
}
\arguments{
\item{Bucket}{[required]}

\item{ContentMD5}{}

\item{LifecycleConfiguration}{}
}
\description{
Deprecated, see the PutBucketLifecycleConfiguration operation.
}
\examples{

}
