% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.s3_operations.R
\name{put_bucket_metrics_configuration}
\alias{put_bucket_metrics_configuration}
\title{Sets a metrics configuration (specified by the metrics configuration ID) for the bucket}
\usage{
put_bucket_metrics_configuration(
  Bucket = "string",
  Id = "string",
  MetricsConfiguration = list(
    Id = "string",
    Filter = list(
      Prefix = "string",
      Tag = list(
        Key = "string",
        Value = "string"
      ),
      And = list(
        Prefix = "string",
        Tags = list(
          list(
            Key = "string",
            Value = "string"
          )
        )
      )
    )
  )
)
}
\arguments{
\item{Bucket}{[required] The name of the bucket for which the metrics configuration is set.}

\item{Id}{[required] The ID used to identify the metrics configuration.}

\item{MetricsConfiguration}{[required] Specifies the metrics configuration.}
}
\description{
Sets a metrics configuration (specified by the metrics configuration ID) for the bucket.
}
\examples{

}
