% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.s3_operations.R
\name{put_bucket_policy}
\alias{put_bucket_policy}
\title{Replaces a policy on a bucket}
\usage{
put_bucket_policy(Bucket, ContentMD5 = NULL,
  ConfirmRemoveSelfBucketAccess = NULL, Policy)
}
\arguments{
\item{Bucket}{[required]}

\item{ContentMD5}{}

\item{ConfirmRemoveSelfBucketAccess}{Set this parameter to true to confirm that you want to remove your permissions to change this bucket policy in the future.}

\item{Policy}{[required] The bucket policy as a JSON document.}
}
\description{
Replaces a policy on a bucket. If the bucket already has a policy, the one in this request completely replaces it.
}
\section{Accepted Parameters}{
\preformatted{put_bucket_policy(
  Bucket = "string",
  ContentMD5 = "string",
  ConfirmRemoveSelfBucketAccess = TRUE|FALSE,
  Policy = "string"
)
}
}

\examples{
# The following example sets a permission policy on a bucket.
\donttest{put_bucket_policy(
  Bucket = "examplebucket",
  Policy = "{\\"Version\\": \\"2012-10-17\\", \\"Statement\\": [{ \\"Sid\\": \\"id-1\\",\\"Effect\\": \\"Allow\\",\\"Principal\\": {\\"AWS\\": \\"arn:aws:iam::123456789012:root\\"}, \\"Action\\": [ \\"s3:PutObject\\",\\"s3:PutObjectAcl\\"], \\"Resource\\": [\\"arn:aws:s3:::acl3/*\\" ] } ]}"
)}

}
