% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.s3_operations.R
\name{put_bucket_tagging}
\alias{put_bucket_tagging}
\title{Sets the tags for a bucket}
\usage{
put_bucket_tagging(
  Bucket = "string",
  ContentMD5 = "string",
  Tagging = list(
    TagSet = list(
      list(
        Key = "string",
        Value = "string"
      )
    )
  )
)
}
\arguments{
\item{Bucket}{[required]}

\item{ContentMD5}{}

\item{Tagging}{[required]}
}
\description{
Sets the tags for a bucket.
}
\examples{
# The following example sets tags on a bucket. Any existing tags are
# replaced.
put_bucket_tagging(
  Bucket = "examplebucket",
  Tagging = list(
    TagSet = list(
      list(
        Key = "Key1",
        Value = "Value1"
      ),
      list(
        Key = "Key2",
        Value = "Value2"
      )
    )
  )
)

}
