% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.s3_operations.R
\name{put_bucket_versioning}
\alias{put_bucket_versioning}
\title{Sets the versioning state of an existing bucket}
\usage{
put_bucket_versioning(
  Bucket = "string",
  ContentMD5 = "string",
  MFA = "string",
  VersioningConfiguration = list(
    MFADelete = "Enabled"|"Disabled",
    Status = "Enabled"|"Suspended"
  )
)
}
\arguments{
\item{Bucket}{[required]}

\item{ContentMD5}{}

\item{MFA}{The concatenation of the authentication device's serial number, a space, and the value that is displayed on your authentication device.}

\item{VersioningConfiguration}{[required]}
}
\description{
Sets the versioning state of an existing bucket. To set the versioning state, you must be the bucket owner.
}
\examples{
# The following example sets versioning configuration on bucket. The
# configuration enables versioning on the bucket.
put_bucket_versioning(
  Bucket = "examplebucket",
  VersioningConfiguration = list(
    MFADelete = "Disabled",
    Status = "Enabled"
  )
)

}
