% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.s3_operations.R
\name{put_object_lock_configuration}
\alias{put_object_lock_configuration}
\title{Places an Object Lock configuration on the specified bucket}
\usage{
put_object_lock_configuration(
  Bucket = "string",
  ObjectLockConfiguration = list(
    ObjectLockEnabled = "Enabled",
    Rule = list(
      DefaultRetention = list(
        Mode = "GOVERNANCE"|"COMPLIANCE",
        Days = 123,
        Years = 123
      )
    )
  ),
  RequestPayer = "requester",
  Token = "string",
  ContentMD5 = "string"
)
}
\arguments{
\item{Bucket}{[required] The bucket whose Object Lock configuration you want to create or replace.}

\item{ObjectLockConfiguration}{The Object Lock configuration that you want to apply to the specified bucket.}

\item{RequestPayer}{}

\item{Token}{}

\item{ContentMD5}{The MD5 hash for the request body.}
}
\description{
Places an Object Lock configuration on the specified bucket. The rule specified in the Object Lock configuration will be applied by default to every new object placed in the specified bucket.
}
\examples{

}
