% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.s3_operations.R
\name{put_public_access_block}
\alias{put_public_access_block}
\title{Creates or modifies the \code{PublicAccessBlock} configuration for an Amazon S3 bucket}
\usage{
put_public_access_block(
  Bucket = "string",
  ContentMD5 = "string",
  PublicAccessBlockConfiguration = list(
    BlockPublicAcls = TRUE|FALSE,
    IgnorePublicAcls = TRUE|FALSE,
    BlockPublicPolicy = TRUE|FALSE,
    RestrictPublicBuckets = TRUE|FALSE
  )
)
}
\arguments{
\item{Bucket}{[required] The name of the Amazon S3 bucket whose \code{PublicAccessBlock} configuration you want to set.}

\item{ContentMD5}{The MD5 hash of the \code{PutPublicAccessBlock} request body.}

\item{PublicAccessBlockConfiguration}{[required] The \code{PublicAccessBlock} configuration that you want to apply to this Amazon S3 bucket. You can enable the configuration options in any combination. For more information about when Amazon S3 considers a bucket or object public, see \href{https://docs.aws.amazon.com/AmazonS3/latest/dev/access-control-block-public-access.html#access-control-block-public-access-policy-status}{The Meaning of "Public"} in the \emph{Amazon Simple Storage Service Developer Guide}.}
}
\description{
Creates or modifies the \code{PublicAccessBlock} configuration for an Amazon S3 bucket.
}
\examples{

}
