% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.s3_operations.R
\name{delete_bucket_website}
\alias{delete_bucket_website}
\title{This operation removes the website configuration from the bucket}
\usage{
delete_bucket_website(Bucket)
}
\arguments{
\item{Bucket}{[required]}
}
\description{
This operation removes the website configuration from the bucket.
}
\section{Accepted Parameters}{
\preformatted{delete_bucket_website(
  Bucket = "string"
)
}
}

\examples{
# The following example deletes bucket website configuration.
\donttest{delete_bucket_website(
  Bucket = "examplebucket"
)}

}
