% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.s3_operations.R
\name{get_bucket_cors}
\alias{get_bucket_cors}
\title{Returns the cors configuration for the bucket}
\usage{
get_bucket_cors(Bucket)
}
\arguments{
\item{Bucket}{[required]}
}
\description{
Returns the cors configuration for the bucket.
}
\section{Accepted Parameters}{
\preformatted{get_bucket_cors(
  Bucket = "string"
)
}
}

\examples{
# The following example returns cross-origin resource sharing (CORS)
# configuration set on a bucket.
\donttest{get_bucket_cors(
  Bucket = "examplebucket"
)}

}
