% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.s3_operations.R
\name{get_bucket_lifecycle}
\alias{get_bucket_lifecycle}
\title{Deprecated, see the GetBucketLifecycleConfiguration operation}
\usage{
get_bucket_lifecycle(
  Bucket = "string"
)
}
\arguments{
\item{Bucket}{[required]}
}
\description{
Deprecated, see the GetBucketLifecycleConfiguration operation.
}
\examples{
# The following example gets ACL on the specified bucket.
get_bucket_lifecycle(
  Bucket = "acl1"
)

}
