% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.s3_operations.R
\name{get_bucket_tagging}
\alias{get_bucket_tagging}
\title{Returns the tag set associated with the bucket}
\usage{
get_bucket_tagging(Bucket)
}
\arguments{
\item{Bucket}{[required]}
}
\description{
Returns the tag set associated with the bucket.
}
\section{Accepted Parameters}{
\preformatted{get_bucket_tagging(
  Bucket = "string"
)
}
}

\examples{
# The following example returns tag set associated with a bucket
\donttest{get_bucket_tagging(
  Bucket = "examplebucket"
)}

}
