% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.s3_operations.R
\name{get_bucket_website}
\alias{get_bucket_website}
\title{Returns the website configuration for a bucket}
\usage{
get_bucket_website(Bucket)
}
\arguments{
\item{Bucket}{[required]}
}
\description{
Returns the website configuration for a bucket.
}
\section{Accepted Parameters}{
\preformatted{get_bucket_website(
  Bucket = "string"
)
}
}

\examples{
# The following example retrieves website configuration of a bucket.
\donttest{get_bucket_website(
  Bucket = "examplebucket"
)}

}
