% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.s3_operations.R
\name{get_object_tagging}
\alias{get_object_tagging}
\title{Returns the tag-set of an object}
\usage{
get_object_tagging(Bucket, Key, VersionId = NULL)
}
\arguments{
\item{Bucket}{[required]}

\item{Key}{[required]}

\item{VersionId}{}
}
\description{
Returns the tag-set of an object.
}
\section{Accepted Parameters}{
\preformatted{get_object_tagging(
  Bucket = "string",
  Key = "string",
  VersionId = "string"
)
}
}

\examples{
# The following example retrieves tag set of an object.
\donttest{get_object_tagging(
  Bucket = "examplebucket",
  Key = "HappyFace.jpg"
)}

# The following example retrieves tag set of an object. The request
# specifies object version.
\donttest{get_object_tagging(
  Bucket = "examplebucket",
  Key = "exampleobject",
  VersionId = "ydlaNkwWm0SfKJR.T1b1fIdPRbldTYRI"
)}

}
